Changed dnd bestiary compatability to 5e tools

+++ index.php
@@ -3,19 +3,23 @@
 
     if(isset($_GET['action'])) {
         $action = $_GET['action'];
-        if($action == 'adjustHealth') {
+        if($action == 'adjustHealth' && isset($_GET['characterId'])) {
             $characterId = $_GET['characterId'];
             $health = (int)mysqli_fetch_row(mysqli_query($conn, "SELECT health FROM current_fight WHERE id = $characterId"))[0] + (int)$_GET['healthNumber'];
             mysqli_query($conn, "UPDATE current_fight set health = $health WHERE id = $characterId");
-        } elseif($action == 'changeInitiative') {
+        } elseif($action == 'changeInitiative' && isset($_GET['characterId'])) {
             $characterId = $_GET['characterId'];
             $initiative = $_GET['initiative'];
             mysqli_query($conn, "UPDATE current_fight SET initiative = $initiative WHERE id = $characterId");
-        } elseif($action == 'delete') {
+        } elseif($action == 'delete' && isset($_GET['characterId'])) {
             $characterId = $_GET['characterId'];
             mysqli_query($conn, "DELETE FROM current_fight WHERE id = $characterId");
-        } elseif($action = 'deleteAllEnemies') {
+        } elseif($action == 'deleteAllEnemies') {
             mysqli_query($conn, "DELETE FROM current_fight WHERE is_player = 0");
+        } elseif($action == 'changeAC' && isset($_GET['characterId'])) {
+            $characterId = $_GET['characterId'];
+            $AC = $_GET['AC'];
+            mysqli_query($conn, "UPDATE current_fight SET AC = $AC WHERE id = $characterId");
         }
         header("Location: index.php");
     }
@@ -39,8 +43,9 @@
         <input type="number" name="characterId" id="characterIdInput">
         <input type="number" name="healthNumber" id="healthInput">
         <input type="number" name="initiative" id="initiativeInput">
+        <input type="number" name="AC" id="ACInput">
     </form>
-    <div id="moreInfoMenu"><div id="closeMoreInfo">X</div><span id="moreInfo"></span></div>
+    <div id="moreInfoMenu"><div class="closeBtnWrapper"><div id="closeMoreInfo">X</div></div><span id="moreInfo"></span></div>
     <header>Fight</header>
     <main>
         <div class="listOfCharacters" id="listOfCharacters">
@@ -54,6 +59,7 @@
                     $min_health = $result['min_health'];
                     $max_health = $result['max_health'];
                     $initiativeBonus = $result['initiative_bonus'];
+                    $isSurprised = isset($_GET['isSurprised']);
                     $AC = $result['AC'];
                     $enemyId = $result['id'];
                     $counter = 0;
@@ -61,6 +67,12 @@
                     for($i = 0; $i < $enemyQuantity; $i++) {
                         $health = rand($min_health, $max_health);
                         $initiative = rand(1, 20) + $initiativeBonus;
+                        if($isSurprised) {
+                            $secondInititative = rand(1, 20) + $initiativeBonus;
+                            if($secondInititative < $initiative) {
+                                $initiative = $secondInititative;
+                            }
+                        }
                         $sql = "INSERT INTO current_fight(name, health, initiative, AC, is_player, enemy_id) Values('$enemyType".$enemyNumber+$counter."', $health, $initiative, $AC, 0, $enemyId)";
                         $result = mysqli_query($conn, $sql);
                         $counter++;
@@ -74,17 +86,16 @@
                 while($row = mysqli_fetch_assoc($result)) {
                     echo "<div class='character' data-characterId='".$row['id']."'>";
                     if($row['enemy_id'] != null) {
-                        $moreInfo = mysqli_fetch_assoc(mysqli_query($conn, "SELECT more_info FROM enemies WHERE id = ".$row['enemy_id']));
-                        echo "<span style='display: none;' class='moreInfo'>".$moreInfo['more_info']."</span>";
+                        $moreInfo = mysqli_fetch_assoc(mysqli_query($conn, "SELECT name, more_info FROM enemies WHERE id = ".$row['enemy_id']));
+                        echo "<span style='display: none;' class='name'><h1>".$moreInfo['name']."</h1></span>";
+                        echo "<div style='display: none;' class='moreInfo'>".$moreInfo['more_info']."</div>";
                     }
                     echo "<span class='characterName'>".$row['name']."</span>";
                     echo "<span class='characterHealth'>Health: ".$row['health']."</span>";
-                    echo "<span>Initiative: </span><input class='no-spinner' type='number' value=".$row['initiative']." id='modifiedInitiativeInput'></input>";
-                    echo "<span>AC: ".$row['AC']."</span>";
-                    echo '<button class="redBtn">sub</button><input class="no-spinner" type="number" id="healthInput"></input><button class="greenBtn">add</button>';
-                    if($row['is_player'] == 1) {
-                        echo "<span style='color: var(--secondary)'>X</span>";
-                    } else {
+                    echo "<span style='display: flex; gap: 10px; align-items: center;'>Initiative: <input class='no-spinner' type='number' value=".$row['initiative']." id='modifiedInitiativeInput'></input></span>";
+                    echo "<span style='display: flex; gap: 10px; align-items: center;'>AC: <input class='no-spinner' type='number' value=".$row['AC']." id='modifiedACInput'></input></span>";
+                    echo '<span class="btnsSurroundingInput"><button class="redBtn">sub</button><input class="no-spinner" type="number" id="healthInput"></input><button class="greenBtn">add</button>';
+                    if($row['is_player'] != 1) {
                         echo "<span class='deleteBtn'>X</span>";
                     }
                     echo "</div>";
@@ -94,7 +105,7 @@
         <form action="" method="get" class="addEnemy">
             <select id="enemySelect" name="enemyType">
                 <?php
-                    $sql = "SELECT * FROM enemies";
+                    $sql = "SELECT * FROM enemies order by name asc";
                     $result = mysqli_query($conn, $sql);
                     while($row = mysqli_fetch_assoc($result)) {
                         echo "<option value='".$row['name']."'>";
@@ -103,7 +114,10 @@
                     }
                 ?>
             </select>
-            <input type="number" name="enemyQuantity" value="1">
+            <div class="row">
+                <input style="border: 0;" type="number" name="enemyQuantity" value="1" max="100">
+                <div class="checkboxInput">Suprised: <input type="checkbox" name="isSurprised"></div>
+            </div>
             <button>Add Enemy</button>
             <button type="button" id="deleteEnemiesBtn">Delete all enemies</button>
             <button type="button" id="addAnotherEnemyBtn">Add another enemy</button>
